test: fold copy-pasted per-encoding tests into enc_* templates
Many per-encoding tests (bzip, gzip, lzma_xz, flac) are variants of the
same test body.  Where the copies are equivalent, replace them with a
shared enc_* template plus a thin per-encoding wrapper supplying
ENC_SUFFIX, ENC_ENCODING and the external ENC_COMPRESS command,
following the existing enc_nframes.c convention.

For tests in this (test X encoding) matrix that already existed in
cut/paste form, conversion was verified to be identical by comparing the
preprocessed output of the wrapper against the original.

Because the (test X encoding) matrix was not fully populated, there is
also a net increase of test coverage as a result of this commit.
