+++ test/flac_get_get2.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2015, 2017 D. V. Wiebe
+/* Copyright (C) 2026 G. Smecher
  *
  ***************************************************************************
  *
@@ -20,62 +20,19 @@
  */
 #include "test.h"
 
-int main(void)
-{
-#if !defined USE_FLAC || !defined TEST_FLAC
-  return 77; /* skip test */
-#else
-  const char *filedir = "dirfile";
-  const char *format = "dirfile/format";
-  const char *data = "dirfile/data";
-  const char *flacdata = "dirfile/data.flac";
-  uint16_t c1[8], c2[8];
-  char command[4096];
-  int i, n1, error1, n2, error2, r = 0;
-  DIRFILE *D;
-
-  memset(c1, 0, 16);
-  memset(c2, 0, 16);
-  rmdirfile();
-  mkdir(filedir, 0700);
-
-  MAKEFORMATFILE(format, "data RAW UINT16 8\n");
-  MAKEDATAFILE(data, uint16_t, i, 256);
-
-#ifdef WORDS_BIGENDIAN
-#define ENDIANNESS "--endian=big"
-#else
-#define ENDIANNESS "--endian=little"
+#ifndef TEST_FLAC
+#define ENC_SKIP_TEST 1
 #endif
 
-  /* compress */
-  snprintf(command, 4096,
-      "\"%s\" " ENDIANNESS " --silent --sample-rate=1 --channels=1 --bps=16 "
-      "--sign=signed --delete-input-file %s >%s 2>%s", FLAC,
-      data, NULL_DEVICE, NULL_DEVICE);
-  if (gd_system(command))
-    return 1;
-
-  D = gd_open(filedir, GD_RDONLY | GD_VERBOSE);
-  n1 = gd_getdata(D, "data", 0, 0, 1, 0, GD_UINT16, c1);
-  error1 = gd_error(D);
-  n2 = gd_getdata(D, "data", 0, 0, 1, 0, GD_UINT16, c2);
-  error2 = gd_error(D);
-  gd_discard(D);
-
-  unlink(flacdata);
-  unlink(format);
-  rmdir(filedir);
-
-  CHECKI(error1, 0);
-  CHECKI(error2, 0);
-  CHECKI(n1, 8);
-  CHECKI(n2, 8);
-  for (i = 0; i < 8; ++i) {
-    CHECKUi(i,c1[i], i);
-    CHECKUi(i,c2[i], i);
-  }
-
-  return r;
+#ifdef USE_FLAC
+#define USE_ENC 1
 #endif
-}
+
+#define ENC_SUFFIX ".flac"
+#define ENC_COMPRESS \
+  snprintf(command, 4096, \
+  "\"%s\" " ENDIANNESS " --silent --sample-rate=1 --channels=1 --bps=16 " \
+  "--sign=signed --delete-input-file %s >%s 2>%s", FLAC, \
+  data, NULL_DEVICE, NULL_DEVICE)
+
+#include "enc_get_get2.c"
