+++ test/lzma_xz_put_endian.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2014, 2017 D.V. Wiebe
+/* Copyright (C) 2026 G. Smecher
  *
  ***************************************************************************
  *
@@ -20,95 +20,17 @@
  */
 #include "test.h"
 
-#ifdef WORDS_BIGENDIAN
-#define ENDIAN "ENDIAN little"
-#else
-#define ENDIAN "ENDIAN big"
-#endif
-
-int main(void)
-{
 #ifndef TEST_LZMA
-  return 77;
-#else
-  const char *filedir = "dirfile";
-  const char *format = "dirfile/format";
-  const char *data_xz = "dirfile/data.xz";
-  const char *data = "dirfile/data";
-  uint16_t c[8];
-#ifdef USE_LZMA
-  char command[4096];
-  uint16_t d;
-#endif
-  struct stat buf;
-  int fd, i, n, e1, e2, stat_data, unlink_data, r = 0;
-  DIRFILE *D;
-
-  rmdirfile();
-  mkdir(filedir, 0700);
-
-  for (i = 0; i < 8; ++i)
-    c[i] = (uint16_t)(0x102 * i);
-
-  MAKEFORMATFILE(format, "data RAW UINT16 8\n" ENDIAN "\n");
-
-#ifdef USE_LZMA
-  D = gd_open(filedir, GD_RDWR | GD_LZMA_ENCODED | GD_VERBOSE);
-#else
-  D = gd_open(filedir, GD_RDWR | GD_LZMA_ENCODED);
-#endif
-  n = gd_putdata(D, "data", 5, 0, 1, 0, GD_UINT16, c);
-  e1 = gd_error(D);
-
-  e2 = gd_close(D);
-  CHECKI(e2, 0);
-
-  stat_data = stat(data_xz, &buf);
-#ifdef USE_LZMA
-  if (stat_data) {
-    perror("stat");
-  }
-  CHECKI(stat_data, 0);
-#else
-  CHECKI(stat_data, -1);
+#define ENC_SKIP_TEST 1
 #endif
 
 #ifdef USE_LZMA
-  /* uncompress */
-  snprintf(command, 4096, "\"%s\" --decompress -f %s > %s", XZ, data_xz, NULL_DEVICE);
-  if (gd_system(command)) {
-    r = 1;
-  } else {
-    fd = open(data, O_RDONLY | O_BINARY);
-    if (fd >= 0) {
-      i = 0;
-      while (read(fd, &d, sizeof(uint16_t))) {
-        if (i < 40 || i > 48) {
-          CHECKIi(i, d, 0);
-        } else
-          CHECKIi(i, d, 0x201 * (i - 40));
-        i++;
-      }
-      CHECKI(i, 48);
-      close(fd);
-    }
-  }
+#define USE_ENC 1
 #endif
 
-  unlink_data = unlink(data);
-  unlink(format);
-  rmdir(filedir);
+#define ENC_SUFFIX ".xz"
+#define ENC_ENCODING GD_LZMA_ENCODED
+#define ENC_COMPRESS \
+  snprintf(command, 4096, "\"%s\" --decompress -f %s > %s", XZ, encdata, NULL_DEVICE)
 
-#ifdef USE_LZMA
-  CHECKI(unlink_data, 0);
-  CHECKI(e1, GD_E_OK);
-  CHECKI(n, 8);
-#else
-  CHECKI(unlink_data, -1);
-  CHECKI(e1, GD_E_UNSUPPORTED);
-  CHECKI(n, 0);
-#endif
-
-  return r;
-#endif
-}
+#include "enc_put_endian.c"
